Apply game logic systems only while Flapping
Here we’re updating the app setup so that we can cycle between the different game states instead of exiting the app when the game is over. The game logic like move_walls and hit_wall is now only applied when the game is in the Flapping GamePhase. When the game is over, we now transition into the GameOver phase instead of exiting the app. This will allow us to show the correct menus at the correct time.