Add a loading screen to show on app launch
This adds a loading screen using the egui crate, to be shown to the user while the menu assets are loaded asynchronously by bevy. We’ve added another GamePhase called Loading and set that to the default state. We then add a setup and exit system to the app for this new phase, so that the loading screen is shown at the correct time.