Implement tweening and rotation for smoother performance
This adds a struct called PhysicsPosition, which tracks the start and end positon of each frame. Knowing the ending position before we render it, means we can smoothly interpolate between the start and end frames. We can also rotate the dragon to face the direction of travel, by calculating the angle between the start and end position.